refactor!: rename module
The only public API should be the CLI (in `__main__.py`).

BREAKNG CHANGE: module `aserehe.check` is no longer in public API.
